refactor: extract plugin dispatch and helpers out of lib.rs
Splits gateway-core/src/lib.rs (was 1055 lines) into focused modules:

- ctx/mod.rs        — GatewayCtx struct
- headers/mod.rs    — HeaderEdit trait, impls, apply_header_modifications,
                      headers_hashmap_to_http_headermap
- proxy_utils/mod.rs — call_interceptor, call_interceptor_blocking,
                       merge_options, js_body_from_content_type,
                       js_body_to_bytes
- upstream_plugin/mod.rs — dispatch() entry point encapsulating the full
                            plugin request lifecycle (steps A–E)

lib.rs is now a thin router (~550 lines): route match → on_request phase 1
→ upstream_plugin::dispatch() for plugin routes, or fall through to the
Pingora pipeline for HTTP routes.

Pure structural refactoring — no behaviour changes.

Closes #46
